refactor: move report-only code to scripts and remove machinery tests
- Move ClusterComparison, AprioriComparison, and 9 diagnostic methods to scripts/report/ as lightweight standalone functions returning plain dicts
- Delete 6 production files whose contents were entirely report-only (comparison.py, apriori.py, association/schemas.py, reduction/schemas.py, and their test files)
- Remove 11 methods from production classes (gtest_matrix, compare_measures, communities, diagnostics, trade_alignment, cophenetic_comparison, dendrogram_data, loadings, alignment, skill_to_cluster, statistics, _log_density)
- Remove 9 Pydantic schemas that only served report-only methods (AprioriResult, CommunityLabel, GraphDiagnostics, MeasureComparison, ComparisonResult, CopheneticResult, ComponentLoading, AlignmentDiagnostics, CorpusStatistics)
- Remove tests that confirm library guarantees rather than production behavior (scipy linkage shape, B^T@B symmetry, Louvain coverage, StandardScaler unit variance, DictVectorizer column consistency)
- Remove tests that confirm Pydantic validates correctly (extra="forbid" rejection, Field(ge=0) bounds, required field enforcement)
- Clean up section headers in test files to only appear with 3+ grouped tests
